requirements.txt
Flask>=3.0.1
tensoflow
Pillow
pandas
